# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Tasks and swagger api.",
        "title": "Cybertask API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": ":7730",
    "basePath": "/",
    "paths": {
        "/task": {
            "put": {
                "description": "Update task via replacing it with provided task in body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Update task.",
                "operationId": "upadte",
                "parameters": [
                    {
                        "description": "update task",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create provided in body.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Create task.",
                "operationId": "create",
                "parameters": [
                    {
                        "description": "create task",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    }
                }
            }
        },
        "/task/:id": {
            "get": {
                "description": "Get task if exist with ID provided in query params.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Get task.",
                "operationId": "get",
                "parameters": [
                    {
                        "description": "get task",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.GetTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete task if exist with ID provided in query params.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "task"
                ],
                "summary": "Delete task.",
                "operationId": "delete",
                "parameters": [
                    {
                        "description": "delete task",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DeleteTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.TaskError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.CreateTaskRequest": {
            "type": "object",
            "properties": {
                "task": {
                    "$ref": "#/definitions/model.Task"
                }
            }
        },
        "handler.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.DeleteTaskRequest": {
            "type": "object"
        },
        "handler.DeleteTaskResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.GetTaskRequest": {
            "type": "object"
        },
        "handler.GetTaskResponse": {
            "type": "object",
            "properties": {
                "task": {
                    "$ref": "#/definitions/model.Task"
                }
            }
        },
        "handler.TaskError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateTaskRequest": {
            "type": "object",
            "properties": {
                "task": {
                    "$ref": "#/definitions/model.Task"
                }
            }
        },
        "handler.UpdateTaskResponse": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.Task": {
            "type": "object",
            "required": [
                "created_at"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "header": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}